usePeriodsを作成
